Добавил базовые проверки соединения с Supabase и Qdrant и чек наличия коллекции в Qdrant

+++ scripts/11_insert_to_qdrant.py
@@ -45,6 +45,10 @@ class QdrantInsertor:
         """
         Вставляет чанки в Qdrant.
         """
+        if not self.check_collection(collection_name=collection_name):
+            print(f"Коллекция {collection_name} не существует")
+            return
+
         self.qdrant_client.upsert(
             collection_name=collection_name,
             points=chunks,
@@ -59,8 +63,13 @@ class QdrantInsertor:
         distance: models.Distance = models.Distance.COSINE,
     ) -> None:
         """
-        Создает коллекцию в Qdrant.
+        Создает коллекцию в Qdrant. 
+        Если коллекция уже существует, то ничего не делает.
         """
+        if self.check_collection(collection_name=collection_name):
+            print(f"Коллекция {collection_name} уже существует")
+            return
+
         self.qdrant_client.create_collection(
             collection_name=collection_name,
             vectors_config=models.VectorParams(
@@ -76,7 +85,22 @@ class QdrantInsertor:
         collection_name: str,
     ) -> None:
         """
-        Удаляет коллекцию в Qdrant.
+        Удаляет коллекцию в Qdrant. Если коллекция не существует, то ничего не делает.
         """
+        if not self.check_collection(collection_name=collection_name):
+            print(f"Коллекция {collection_name} не существует")
+            return
+
         self.qdrant_client.delete_collection(collection_name=collection_name)
         print(f"Коллекция {collection_name} удалена")
+
+    def check_collection(
+        self,
+        *,
+        collection_name: str,
+    ) -> bool:
+        """
+        Проверяет наличие коллекции в Qdrant.
+        True - коллекция существует, False - коллекция не существует.
+        """
+        return self.qdrant_client.collection_exists(collection_name=collection_name)
